backend/requirements.txt
fastapi
uvicorn
pydantic
pandas
sqlalchemy
psycopg2-binary
alembic
bcrypt
python-jose[cryptography]
python-multipart
